+++ kh7finalteam1/src/main/java/com/kh/finalteam1/controller/LoginController.java
@@ -5,41 +5,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class LoginController {
-@GetMapping("/login")
-public String login() {
-	return "login/loginstart"; 
-}
-@Controller
-public class Join1Controller{
-	@GetMapping("/join1")
-	public String login() {
-		return "login/join1";
-	}
-	
 	@GetMapping("/login")
-	public String login1() {
-		return "login/login"; 
-	}
-}
-@Controller
-public class Join2Controller{
-	@GetMapping("/join2")
 	public String login() {
-		return "login/join2";
+		return "login/loginstart"; 
 	}
-}
-@Controller
-public class Join3Controller{
-	@GetMapping("join3")
-	public String login() {
-		return "login/join3";
-	}
-}
-@Controller
-public class Join4Controller{
-	@GetMapping("join4")
-	public String login() {
-		return "login/join4";
-	}
-}
+
+
 }
