+++ kh7finalteam1/src/main/webapp/WEB-INF/views/template/adminSidebar.jsp
@@ -69,7 +69,7 @@
                     <div class="bg-white py-2 collapse-inner rounded">
                         <h6 class="collapse-header">Content:</h6>
                         <a class="collapse-item" href="${pageContext.request.contextPath}/admin/content/">컨텐츠 리스트</a>
-                        <a class="collapse-item" href="${pageContext.request.contextPath}/admin/contentRegist">컨텐츠 추가</a>
+                        <a class="collapse-item" href="${pageContext.request.contextPath}/admin/content/contentRegist">컨텐츠 추가</a>
                         <a class="collapse-item" href="${pageContext.request.contextPath}/admin/genre">장르 추가</a>
                         <a class="collapse-item" href="${pageContext.request.contextPath}/admin/feature">특징 추가</a>
                         <a class="collapse-item" href="forgot-password.html">배우 추가</a>
@@ -183,7 +183,7 @@
             <!-- Sidebar Message   -->
             <div class="sidebar-card d-none d-lg-flex">
                 <div style="padding:1em 2em 1em 2em"><img class="sidebar-card-illustration mb-2" src="${pageContext.request.contextPath }/res/img/weflex_logo.png" alt="..." style="width: 100%;"></div>
-                <p class="text-center mb-2"><strong>넷플릭스</strong>는 TV 프로그램과 영화를 인터넷 연결 지원 디바이스에서 광고 없이 시청할 수 있는 멤버십 기반 스트리밍 서비스입니다.</p>
+                <p class="text-center mb-2"><strong>위플랙스</strong>는 TV 프로그램과 영화를 인터넷 연결 지원 디바이스에서 광고 없이 시청할 수 있는 멤버십 기반 스트리밍 서비스입니다.</p>
                 <a class="btn btn-success btn-sm" href="https://www.netflix.com/browse">Go! Weflex</a>
             </div>
